Move controllers and stores to versioned 'v2' packages
