+++ pkg/controller/transaction/v2/watcher.go
@@ -2,12 +2,12 @@
 //
 // SPDX-License-Identifier: Apache-2.0
 
-package proposal
+package v2
 
 import (
 	"context"
-	configurationstore "github.com/onosproject/onos-config/pkg/store/configuration"
-	proposalstore "github.com/onosproject/onos-config/pkg/store/proposal"
+	proposalstore "github.com/onosproject/onos-config/pkg/store/proposal/v2"
+	transactionstore "github.com/onosproject/onos-config/pkg/store/transaction/v2"
 	"sync"
 
 	configapi "github.com/onosproject/onos-api/go/onos/config/v2"
@@ -17,15 +17,57 @@ import (
 
 const queueSize = 100
 
-// Watcher proposal store watcher
+// Watcher transaction store watcher
 type Watcher struct {
+	transactions transactionstore.Store
+	cancel       context.CancelFunc
+	mu           sync.Mutex
+}
+
+// Start starts the watcher
+func (w *Watcher) Start(ch chan<- controller.ID) error {
+	w.mu.Lock()
+	defer w.mu.Unlock()
+	if w.cancel != nil {
+		return nil
+	}
+
+	eventCh := make(chan configapi.TransactionEvent, queueSize)
+	ctx, cancel := context.WithCancel(context.Background())
+
+	err := w.transactions.Watch(ctx, eventCh, transactionstore.WithReplay())
+	if err != nil {
+		cancel()
+		return err
+	}
+	w.cancel = cancel
+	go func() {
+		for event := range eventCh {
+			ch <- controller.NewID(event.Transaction.Index)
+		}
+	}()
+	return nil
+}
+
+// Stop stops the watcher
+func (w *Watcher) Stop() {
+	w.mu.Lock()
+	if w.cancel != nil {
+		w.cancel()
+		w.cancel = nil
+	}
+	w.mu.Unlock()
+}
+
+// ProposalWatcher proposal store watcher
+type ProposalWatcher struct {
 	proposals proposalstore.Store
 	cancel    context.CancelFunc
 	mu        sync.Mutex
 }
 
 // Start starts the watcher
-func (w *Watcher) Start(ch chan<- controller.ID) error {
+func (w *ProposalWatcher) Start(ch chan<- controller.ID) error {
 	w.mu.Lock()
 	defer w.mu.Unlock()
 	if w.cancel != nil {
@@ -43,57 +85,14 @@ func (w *Watcher) Start(ch chan<- controller.ID) error {
 	w.cancel = cancel
 	go func() {
 		for event := range eventCh {
-			ch <- controller.NewID(event.Proposal.ID)
+			ch <- controller.NewID(event.Proposal.TransactionIndex)
 		}
 	}()
 	return nil
 }
 
 // Stop stops the watcher
-func (w *Watcher) Stop() {
-	w.mu.Lock()
-	if w.cancel != nil {
-		w.cancel()
-		w.cancel = nil
-	}
-	w.mu.Unlock()
-}
-
-// ConfigurationWatcher configuration store watcher
-type ConfigurationWatcher struct {
-	configurations configurationstore.Store
-	cancel         context.CancelFunc
-	mu             sync.Mutex
-}
-
-// Start starts the watcher
-func (w *ConfigurationWatcher) Start(ch chan<- controller.ID) error {
-	w.mu.Lock()
-	defer w.mu.Unlock()
-	if w.cancel != nil {
-		return nil
-	}
-
-	eventCh := make(chan configapi.ConfigurationEvent, queueSize)
-	ctx, cancel := context.WithCancel(context.Background())
-
-	err := w.configurations.Watch(ctx, eventCh, configurationstore.WithReplay())
-	if err != nil {
-		cancel()
-		return err
-	}
-	w.cancel = cancel
-	go func() {
-		for event := range eventCh {
-			ch <- controller.NewID(proposalstore.NewID(event.Configuration.TargetID, event.Configuration.Index))
-			ch <- controller.NewID(proposalstore.NewID(event.Configuration.TargetID, event.Configuration.Status.Applied.Index))
-		}
-	}()
-	return nil
-}
-
-// Stop stops the watcher
-func (w *ConfigurationWatcher) Stop() {
+func (w *ProposalWatcher) Stop() {
 	w.mu.Lock()
 	if w.cancel != nil {
 		w.cancel()
