+++ pkg/manager/manager.go
@@ -7,27 +7,27 @@ package manager
 
 import (
 	"github.com/atomix/go-sdk/pkg/client"
-	configurationcontroller "github.com/onosproject/onos-config/pkg/controller/configuration"
+	configurationcontroller "github.com/onosproject/onos-config/pkg/controller/configuration/v2"
 	"github.com/onosproject/onos-config/pkg/controller/connection"
-	mastershipcontroller "github.com/onosproject/onos-config/pkg/controller/mastership"
-	proposalcontroller "github.com/onosproject/onos-config/pkg/controller/proposal"
+	mastershipcontroller "github.com/onosproject/onos-config/pkg/controller/mastership/v2"
+	proposalcontroller "github.com/onosproject/onos-config/pkg/controller/proposal/v2"
 	"github.com/onosproject/onos-config/pkg/controller/target"
-	"github.com/onosproject/onos-config/pkg/store/proposal"
+	proposal "github.com/onosproject/onos-config/pkg/store/proposal/v2"
 
 	"github.com/onosproject/onos-config/pkg/controller/node"
 	"github.com/onosproject/onos-config/pkg/northbound/admin"
 	gnminb "github.com/onosproject/onos-config/pkg/northbound/gnmi/v2"
 	"github.com/onosproject/onos-config/pkg/pluginregistry"
 	sb "github.com/onosproject/onos-config/pkg/southbound/gnmi"
-	"github.com/onosproject/onos-config/pkg/store/configuration"
+	configuration "github.com/onosproject/onos-config/pkg/store/configuration/v2"
 	"github.com/onosproject/onos-config/pkg/store/topo"
-	"github.com/onosproject/onos-config/pkg/store/transaction"
+	transaction "github.com/onosproject/onos-config/pkg/store/transaction/v2"
 	"github.com/onosproject/onos-lib-go/pkg/certs"
 	"github.com/onosproject/onos-lib-go/pkg/northbound"
 
 	"os"
 
-	transactioncontroller "github.com/onosproject/onos-config/pkg/controller/transaction"
+	transactioncontroller "github.com/onosproject/onos-config/pkg/controller/transaction/v2"
 	"github.com/onosproject/onos-lib-go/pkg/logging"
 )
