+++ pkg/northbound/gnmi/v2/get_test.go
@@ -7,11 +7,11 @@ package gnmi
 import (
 	"context"
 	"github.com/atomix/go-sdk/pkg/test"
-	configurationcontroller "github.com/onosproject/onos-config/pkg/controller/configuration"
-	proposalcontroller "github.com/onosproject/onos-config/pkg/controller/proposal"
-	transactioncontroller "github.com/onosproject/onos-config/pkg/controller/transaction"
+	configurationcontroller "github.com/onosproject/onos-config/pkg/controller/configuration/v2"
+	proposalcontroller "github.com/onosproject/onos-config/pkg/controller/proposal/v2"
+	transactioncontroller "github.com/onosproject/onos-config/pkg/controller/transaction/v2"
 	sb "github.com/onosproject/onos-config/pkg/southbound/gnmi"
-	"github.com/onosproject/onos-config/pkg/store/proposal"
+	proposal "github.com/onosproject/onos-config/pkg/store/proposal/v2"
 	"github.com/onosproject/onos-lib-go/pkg/controller"
 	"github.com/openconfig/gnmi/proto/gnmi_ext"
 	"os"
@@ -28,8 +28,8 @@ import (
 	topoapi "github.com/onosproject/onos-api/go/onos/topo"
 	gnmitest "github.com/onosproject/onos-config/pkg/northbound/gnmi/test"
 	plugintest "github.com/onosproject/onos-config/pkg/pluginregistry/test"
-	"github.com/onosproject/onos-config/pkg/store/configuration"
-	"github.com/onosproject/onos-config/pkg/store/transaction"
+	configuration "github.com/onosproject/onos-config/pkg/store/configuration/v2"
+	transaction "github.com/onosproject/onos-config/pkg/store/transaction/v2"
 	"github.com/onosproject/onos-config/pkg/utils"
 	"github.com/onosproject/onos-config/pkg/utils/path"
 	"github.com/openconfig/gnmi/proto/gnmi"
