+++ pkg/store/transaction/v2/store.go
@@ -2,7 +2,7 @@
 //
 // SPDX-License-Identifier: Apache-2.0
 
-package transaction
+package v2
 
 import (
 	"github.com/atomix/go-sdk/pkg/primitive"
